Reuse code between profile stab and product stab tests
